refactor(migrate): extract scaffold.toml migration into its own module
Pull MigrationReport, migrate_to_v0_2_0, and write_repo_ref_via_toml_edit
out of commands/init.rs and split_flake_ref out of config.rs into a new
src/migrate.rs. commands/init.rs shrinks to cmd_init plumbing plus the
end-to-end migration tests; config.rs's doc-comment drops the "migration
helpers" claim and points at crate::migrate instead.

Per review feedback on PR #95.
