+++ src/lib.rs
@@ -6,6 +6,7 @@ pub(crate) mod config;
 pub(crate) mod constants;
 pub(crate) mod doctor_checks;
 pub(crate) mod error;
+pub(crate) mod migrate;
 pub(crate) mod model;
 pub(crate) mod process;
 pub(crate) mod project;
